requirements.txt
accelerate
datasets
evaluate
flask
huggingface_hub
python-dotenv
requests
selenium
torch
torchaudio
torchvision
transformers
webdriver_manager